fixed split
