+++ backend_api/tomato/api/__init__.py
@@ -47,6 +47,6 @@ from resources import resources_map, network_create, network_info, network_insta
 	template_create, template_list, template_modify, template_remove
 
 from topology import topology_action, topology_create, topology_info,\
-	topology_list, topology_modify, topology_permission, topology_permissions, topology_remove, topology_usage
+	topology_list, topology_modify, topology_set_permission, topology_permissions, topology_remove, topology_usage
 
 from orchestration.topology_export_import import topology_export, topology_import
