+++ backend_core/tomato/api/__init__.py
@@ -22,7 +22,7 @@ from capabilities import capabilities, capabilities_connection, capabilities_ele
 from connections import connection_create, connection_info, connection_modify, connection_remove,\
 	connection_action, connection_usage
 
-from debug import debug, debug_stats, debug_services_reachable
+from debug import debug, debug_stats
 
 from docs import docs, role_list
 
@@ -36,10 +36,10 @@ from host import site_create, site_info, site_list, site_modify, site_remove,\
 
 from misc import link_statistics, notifyAdmins, server_info, statistics, task_execute, task_list
 
-from resources import resources_map, network_create, network_info, network_instance_create, network_instance_info,\
+from resources import network_create, network_info, network_instance_create, network_instance_info,\
 	network_instance_list, network_instance_modify, network_instance_remove, network_list, network_modify,\
 	network_remove, profile_create, profile_info, profile_list, profile_modify, profile_remove, template_info,\
 	template_create, template_list, template_modify, template_remove
 
 from topology import topology_action, topology_create, topology_info,\
-	topology_list, topology_modify, topology_permission, topology_permissions, topology_remove, topology_usage
+	topology_list, topology_modify, topology_set_permission, topology_remove, topology_usage
